+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0-alpha.3] - 2025-07-01
+
+- Fixing metadata schema inference
+
+## [0.3.0-alpha.2] - 2025-07-01
+
+- Cleaned up docs to remove schema examples and moved it to examples folder
+
 ## [0.3.0-alpha.1] - 2025-07-01
 
 - Removed `@standard-schema/spec` dependency